Modificaciones en:
+ Buscador en el pueblo, te devuelve un Bicho aleatoreamente (no se
incluye el random todavia).
+ Ahora se registra el campeon de un Duelo en el Dojo.
+ Modificaciones en Tupla.
+ Modificaciones y nuevos test de Buscador.

+++ src/main/java/ar/edu/unq/epers/bichomon/backend/model/buscador/Tupla.java
@@ -6,10 +6,10 @@ import ar.edu.unq.epers.bichomon.backend.model.especie.Especie;
 
 public class Tupla{
 	private Especie key;
-	private double value;
+	private float value;
 	private List<Integer> lsValue;
 
-	public Tupla(Especie especie, double value){
+	public Tupla(Especie especie, float value){
 		this.key 	= especie;
 		this.value 	= value;
 	};
@@ -23,7 +23,7 @@ public class Tupla{
 		return this.key;
 	}
 	
-	public double getValue(){
+	public float getValue(){
 		return this.value;
 	}
